+++ src/main/kotlin/com/challenge/prewave/prewave_challenge/EdgeService.kt
@@ -4,8 +4,6 @@ import com.challenge.prewave.prewave_challenge.api.errors.EdgeAlreadyExistsExcep
 import com.challenge.prewave.prewave_challenge.api.models.Edge
 import com.challenge.prewave.prewave_challenge.tables.references.EDGE
 import org.jooq.DSLContext
-import org.jooq.impl.DSL.field
-import org.jooq.impl.DSL.table
 import org.springframework.dao.DuplicateKeyException
 import org.springframework.stereotype.Service
 
@@ -14,7 +12,7 @@ class EdgeService(private val create: DSLContext) {
 
 
 
-    fun createEdge(fromNode: Long, toNode: Long): Edge {
+    fun createEdge(fromNode: Int, toNode: Int): Edge {
         try {
             create.insertInto(EDGE, EDGE.FROM_ID, EDGE.TO_ID).values(fromNode, toNode).execute()
             return Edge(fromNode = fromNode, toNode = toNode)
