+++ src/main/kotlin/com/challenge/prewave/prewave_challenge/api/models/Edge.kt
@@ -1,3 +1,3 @@
 package com.challenge.prewave.prewave_challenge.api.models
 
-data class Edge(val fromNode: Long, val toNode: Long)
+data class Edge(val fromNode: Int, val toNode: Int)
